Moving scanBarCode API to media namespace
