+++ src/public/media.ts
@@ -23,6 +23,11 @@ export namespace media {
   const mediaAPISupportVersion = '1.8.0';
 
   /**
+   * This is the SDK version when scan barcode API is supported on all three platforms android, iOS and web.
+   */
+  export const scanBarCodeAPIRequiredVersion = '2.0.0';
+
+  /**
    * Enum for file formats supported
    */
   export enum FileFormat {
@@ -405,6 +410,61 @@ export namespace media {
 
     const params = [uriList];
     const messageId = sendMessageRequestToParent('viewImages', params);
+    GlobalVars.callbacks[messageId] = callback;
+  }
+
+  /**
+   * Optional BarCode configuration which can be passed as input to the Scan Bar Code API to customize barcode scanning experience.
+   * All properties in BarCodeConfig are optional and have default values in the platform.
+   */
+  export interface BarCodeConfig {
+    /**
+     * Optional; Lets the developer to specify the type of barcode which can be scanned.
+     * Default value is All i.e Both One Dimensional and Two Dimensional(only QR code) can be scanned
+     */
+    barCodeType?: BarCodeType;
+    /**
+     * Optional; Lets the developer to specify the bar code scan time out interval in seconds.
+     * Default value is 30 seconds and max value is 60 seconds
+     */
+    timeOutIntervalInSec?: number;
+  }
+
+  /**
+   * Specifies the types of barcode which are supported by Barcode API
+   */
+  export enum BarCodeType {
+    /**
+     * Both One dimensional and two dimensional barcode types are supported
+     */
+    All = 1,
+    /**
+     * Only One dimensional barcode type are supported
+     */
+    OneDimensional = 2,
+    /**
+     * Only Two dimensional barcode types are supported. Only QR code will be supported in Phase 1.
+     */
+    TwoDimensional = 3,
+  }
+
+  /**
+   * Scan Barcode/QR code
+   * @param callback callback to invoke after scanning the barcode/QR code
+   * @param config optional input configuration to customize the bar code scanning experience
+   */
+  export function scanBarCode(callback: (error: SdkError, decodedText: string) => void, config?: BarCodeConfig): void {
+    if (!callback) {
+      throw new Error('[media.scanBarCode] Callback cannot be null');
+    }
+    ensureInitialized(FrameContexts.content, FrameContexts.task);
+
+    if (!isAPISupportedByPlatform(scanBarCodeAPIRequiredVersion)) {
+      const oldPlatformError: SdkError = { errorCode: ErrorCode.OLD_PLATFORM };
+      callback(oldPlatformError, undefined);
+      return;
+    }
+    const messageId = sendMessageRequestToParent('media.scanBarCode', [config]);
     GlobalVars.callbacks[messageId] = callback;
   }
 }
